WIP: libusb for all platforms
