+++ src/pc/PlatformData.c
@@ -21,7 +21,6 @@
 #ifndef _WIN32_WINNT
 #define _WIN32_WINNT 0x0601  /* Windows 7. */
 #endif
-#include "win_usb.h"
 #include "win_time.h"
 #include "win_pthread.h"
 #include <winsock2.h>
@@ -32,7 +31,6 @@
 #include <arpa/inet.h>
 #include <netdb.h>
 #include <unistd.h>
-#include <libusb.h>
 typedef int SOCKET;
 #endif
 
@@ -44,45 +42,25 @@ typedef int SOCKET;
 #include <arpa/inet.h>
 #endif  /*USE_LINK_JTAG*/
 
-#define USB_ENDPOINT_IN 0x81
-#define USB_ENDPOINT_OUT 0x01
-
 #ifndef USE_USB_VSC
 #include <sys/wait.h>
 #include <sys/un.h>
 #include <sys/ioctl.h>
 #include <termios.h>
 
+#include "usb_host.h"
+
 extern int usbFdWrite;
 extern int usbFdRead;
 #endif  /*USE_USB_VSC*/
-
-#ifndef XLINK_USB_DATA_TIMEOUT
-#define XLINK_USB_DATA_TIMEOUT 0
-#endif
-
-// ------------------------------------
-// Helpers declaration. Begin.
-// ------------------------------------
-#ifdef USE_USB_VSC
-static int usb_write(libusb_device_handle *f, const void *data, size_t size);
-static int usb_read(libusb_device_handle *f, void *data, size_t size);
-#endif
-// ------------------------------------
-// Helpers declaration. End.
-// ------------------------------------
-
-
 
 // ------------------------------------
 // Wrappers declaration. Begin.
 // ------------------------------------
 
-static int usbPlatformRead(void *fd, void *data, int size);
 static int pciePlatformRead(void *f, void *data, int size);
 static int tcpipPlatformRead(void *fd, void *data, int size);
 
-static int usbPlatformWrite(void *fd, void *data, int size);
 static int pciePlatformWrite(void *f, void *data, int size);
 static int tcpipPlatformWrite(void *fd, void *data, int size);
 
@@ -166,98 +144,6 @@ void XLinkPlatformDeallocateData(void *ptr, uint32_t size, uint32_t alignment)
 // Wrappers implementation. Begin.
 // ------------------------------------
 
-int usbPlatformRead(void* fd, void* data, int size)
-{
-    int rc = 0;
-#ifndef USE_USB_VSC
-    int nread =  0;
-#ifdef USE_LINK_JTAG
-    while (nread < size){
-        nread += read(usbFdWrite, &((char*)data)[nread], size - nread);
-        printf("read %d %d\n", nread, size);
-    }
-#else
-    if(usbFdRead < 0)
-    {
-        return -1;
-    }
-
-    while(nread < size)
-    {
-        int toRead = (PACKET_LENGTH && (size - nread > PACKET_LENGTH)) \
-                        ? PACKET_LENGTH : size - nread;
-
-        while(toRead > 0)
-        {
-            rc = read(usbFdRead, &((char*)data)[nread], toRead);
-            if ( rc < 0)
-            {
-                return -2;
-            }
-            toRead -=rc;
-            nread += rc;
-        }
-        unsigned char acknowledge = 0xEF;
-        int wc = write(usbFdRead, &acknowledge, sizeof(acknowledge));
-        if (wc != sizeof(acknowledge))
-        {
-            return -2;
-        }
-    }
-#endif  /*USE_LINK_JTAG*/
-#else
-    rc = usb_read((libusb_device_handle *) fd, data, size);
-#endif  /*USE_USB_VSC*/
-    return rc;
-}
-
-int usbPlatformWrite(void *fd, void *data, int size)
-{
-    int rc = 0;
-#ifndef USE_USB_VSC
-    int byteCount = 0;
-#ifdef USE_LINK_JTAG
-    while (byteCount < size){
-        byteCount += write(usbFdWrite, &((char*)data)[byteCount], size - byteCount);
-        printf("write %d %d\n", byteCount, size);
-    }
-#else
-    if(usbFdWrite < 0)
-    {
-        return -1;
-    }
-    while(byteCount < size)
-    {
-       int toWrite = (PACKET_LENGTH && (size - byteCount > PACKET_LENGTH)) \
-                        ? PACKET_LENGTH:size - byteCount;
-       int wc = write(usbFdWrite, ((char*)data) + byteCount, toWrite);
-
-       if ( wc != toWrite)
-       {
-           return -2;
-       }
-
-       byteCount += toWrite;
-       unsigned char acknowledge;
-       int rc;
-       rc = read(usbFdWrite, &acknowledge, sizeof(acknowledge));
-
-       if ( rc < 0)
-       {
-           return -2;
-       }
-
-       if (acknowledge != 0xEF)
-       {
-           return -2;
-       }
-    }
-#endif  /*USE_LINK_JTAG*/
-#else
-    rc = usb_write((libusb_device_handle *) fd, data, size);
-#endif  /*USE_USB_VSC*/
-    return rc;
-}
 
 
 #if (defined(_WIN32) || defined(_WIN64))
@@ -424,48 +310,4 @@ static int tcpipPlatformWrite(void *fd, void *data, int size)
 
 // ------------------------------------
 // Wrappers implementation. End.
-// ------------------------------------
-
-
-
-// ------------------------------------
-// Helpers implementation. Begin.
-// ------------------------------------
-#ifdef USE_USB_VSC
-int usb_read(libusb_device_handle *f, void *data, size_t size)
-{
-    const int chunk_size = DEFAULT_CHUNKSZ;
-    while(size > 0)
-    {
-        int bt, ss = (int)size;
-        if(ss > chunk_size)
-            ss = chunk_size;
-        int rc = libusb_bulk_transfer(f, USB_ENDPOINT_IN,(unsigned char *)data, ss, &bt, XLINK_USB_DATA_TIMEOUT);
-        if(rc)
-            return rc;
-        data = ((char *)data) + bt;
-        size -= bt;
-    }
-    return 0;
-}
-
-int usb_write(libusb_device_handle *f, const void *data, size_t size)
-{
-    const int chunk_size = DEFAULT_CHUNKSZ;
-    while(size > 0)
-    {
-        int bt, ss = (int)size;
-        if(ss > chunk_size)
-            ss = chunk_size;
-        int rc = libusb_bulk_transfer(f, USB_ENDPOINT_OUT, (unsigned char *)data, ss, &bt, XLINK_USB_DATA_TIMEOUT);
-        if(rc)
-            return rc;
-        data = (char *)data + bt;
-        size -= bt;
-    }
-    return 0;
-}
-#endif
-// ------------------------------------
-// Helpers implementation. End.
 // ------------------------------------
